Extract buildBuyMetadata helper to remove duplication across buy pages
