+++ apps/client-fnp/app/(landing)/buy-agrochemicals/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 import { queryAgroChemical } from "@/lib/query"
-import { formatProductName } from "@/lib/utilities"
+import { buildBuyMetadata } from "@/lib/utilities"
 import Link from "next/link"
 import { Beaker } from "lucide-react"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
@@ -15,20 +15,7 @@ export async function generateMetadata({ params }: BuyAgroChemicalPageProps): Pr
     const response = await queryAgroChemical(slug).catch(() => null)
     const chemical = response?.data
     if (!chemical) return { title: 'Agrochemical | farmnport.com' }
-    const name = formatProductName(chemical.name)
-    const brand = chemical.brand?.name ? ` ${formatProductName(chemical.brand.name)}` : ''
-    const category = chemical.agrochemical_category?.name || 'Agrochemical'
-    return {
-        title: `${name}${brand} – Buy ${category} Zimbabwe | farmnport.com`,
-        description: chemical.description || `Buy ${name}${brand}. ${category} for Zimbabwe farmers. View dosage rates, pricing and order online at farmnport.com.`,
-        alternates: { canonical: `/buy-agrochemicals/${slug}` },
-        openGraph: {
-            title: `${name}${brand} – ${category} | farmnport.com`,
-            description: chemical.description || `Buy ${name}${brand} online. ${category} for Zimbabwe farmers.`,
-            siteName: 'farmnport',
-            type: 'website',
-        },
-    }
+    return buildBuyMetadata(chemical, chemical.agrochemical_category?.name || 'Agrochemical', `/buy-agrochemicals/${slug}`)
 }
 
 export default async function BuyAgroChemicalPage({ params }: BuyAgroChemicalPageProps) {
