+++ apps/client-fnp/app/(landing)/buy-animal-health/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 import { queryAnimalHealthProduct } from "@/lib/query"
-import { formatProductName } from "@/lib/utilities"
+import { buildBuyMetadata } from "@/lib/utilities"
 import Link from "next/link"
 import { Beaker } from "lucide-react"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
@@ -15,20 +15,7 @@ export async function generateMetadata({ params }: BuyAnimalHealthProductPagePro
     const response = await queryAnimalHealthProduct(slug).catch(() => null)
     const product = response?.data
     if (!product) return { title: 'Animal Health Product | farmnport.com' }
-    const name = formatProductName(product.name)
-    const brand = product.brand?.name ? ` ${formatProductName(product.brand.name)}` : ''
-    const category = product.animal_health_category?.name || 'Animal Health'
-    return {
-        title: `${name}${brand} – Buy ${category} Zimbabwe | farmnport.com`,
-        description: product.description || `Buy ${name}${brand}. ${category} for poultry and livestock in Zimbabwe. View dosage, pricing and order online at farmnport.com.`,
-        alternates: { canonical: `/buy-animal-health/${slug}` },
-        openGraph: {
-            title: `${name}${brand} – ${category} | farmnport.com`,
-            description: product.description || `Buy ${name}${brand} online. ${category} for Zimbabwe farmers.`,
-            siteName: 'farmnport',
-            type: 'website',
-        },
-    }
+    return buildBuyMetadata(product, product.animal_health_category?.name || 'Animal Health', `/buy-animal-health/${slug}`, 'poultry and livestock in Zimbabwe')
 }
 
 export default async function BuyAnimalHealthProductPage({ params }: BuyAnimalHealthProductPageProps) {
