+++ apps/client-fnp/app/(landing)/buy-livestock-poultry/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 import { queryLivestockPoultryProduct } from "@/lib/query"
-import { formatProductName } from "@/lib/utilities"
+import { buildBuyMetadata } from "@/lib/utilities"
 import Link from "next/link"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { BuyProductInteractive } from "@/components/shop/BuyProductInteractive"
@@ -14,20 +14,8 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
     const response = await queryLivestockPoultryProduct(slug).catch(() => null)
     const product = response?.data
     if (!product) return { title: 'Livestock & Poultry | farmnport.com' }
-    const name = formatProductName(product.name)
-    const brand = product.brand?.name ? ` ${formatProductName(product.brand.name)}` : ''
     const category = [product.species, product.type].filter(Boolean).join(' ') || 'Livestock & Poultry'
-    return {
-        title: `${name}${brand} – Buy ${category} Zimbabwe | farmnport.com`,
-        description: product.description || `Buy ${name}${brand}. ${category} for Zimbabwe farmers. View pricing and order online at farmnport.com.`,
-        alternates: { canonical: `/buy-livestock-poultry/${slug}` },
-        openGraph: {
-            title: `${name}${brand} – ${category} | farmnport.com`,
-            description: product.description || `Buy ${name}${brand} online. ${category} available in Zimbabwe.`,
-            siteName: 'farmnport',
-            type: 'website',
-        },
-    }
+    return buildBuyMetadata(product, category, `/buy-livestock-poultry/${slug}`)
 }
 
 export default async function BuyLivestockPoultryProductPage({ params }: Props) {
