+++ apps/client-fnp/app/(landing)/buy-plant-nutrition/[slug]/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next'
 import { queryPlantNutritionProduct } from "@/lib/query"
-import { formatProductName } from "@/lib/utilities"
+import { buildBuyMetadata } from "@/lib/utilities"
 import Link from "next/link"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { BuyProductInteractive } from "@/components/shop/BuyProductInteractive"
@@ -14,20 +14,7 @@ export async function generateMetadata({ params }: BuyPlantNutritionProductPageP
     const response = await queryPlantNutritionProduct(slug).catch(() => null)
     const product = response?.data
     if (!product) return { title: 'Plant Nutrition Product | farmnport.com' }
-    const name = formatProductName(product.name)
-    const brand = product.brand?.name ? ` ${formatProductName(product.brand.name)}` : ''
-    const category = product.plant_nutrition_category?.name || 'Plant Nutrition'
-    return {
-        title: `${name}${brand} – Buy ${category} Zimbabwe | farmnport.com`,
-        description: product.description || `Buy ${name}${brand}. ${category} for Zimbabwe crops. View application rates, pricing and order online at farmnport.com.`,
-        alternates: { canonical: `/buy-plant-nutrition/${slug}` },
-        openGraph: {
-            title: `${name}${brand} – ${category} | farmnport.com`,
-            description: product.description || `Buy ${name}${brand} online. ${category} for Zimbabwe crops.`,
-            siteName: 'farmnport',
-            type: 'website',
-        },
-    }
+    return buildBuyMetadata(product, product.plant_nutrition_category?.name || 'Plant Nutrition', `/buy-plant-nutrition/${slug}`, 'Zimbabwe crops')
 }
 
 export default async function BuyPlantNutritionProductPage({ params }: BuyPlantNutritionProductPageProps) {
